refactoring rtllib
